safari-534.53-branch/LayoutTests/svg/custom/svg-parse-overflow-4.html

<html>
<body>
<script>
document.body.offsetTop;
if (window.layoutTestController)
    layoutTestController.dumpAsText();
</script>
<svg>
<foreignObject y="105.0e50" transform="scale(5) skewX(5)">
</svg>
</body>
