Safari-611.1.10/LayoutTests/svg/custom/svg-parse-overflow-4-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <foreignObject> attribute y="105.0e50"
